Treat `aria-expanded` as a string value when toggling
Both Firefox and Chrome set `aria-expanded="{value}"` even if the
value is boolean and the raw HTML is `aria-expanded={value}`.

Previously, we cast the string to a Boolean value, but while
Firefox casts `'false'` to `false`, Chrome casts `'false'` to
`true`, which made this script only work in Firefox.

Now, we deal only in strings.